Implementation of jump tables complete. went for a dual table solution, one for parent opcodes and another for child opcodes.
Child opcode table is accessed using an offset determined using the childFuncOffset enum that pulls the loNibble to decide which offset to apply.

All opcodes and their functions (and associated pointers) have been implemented.

The source could still use a lot of abstraction and encapsulation work. But for now, I'll begin focusing on main.cpp and cleaning that section up.