Sound Update: Changes to Chip8.cpp and included lib files.
This Chip8 emulator now produces a beeping noise. The noise made can be changed by swapping out the "beep.wav" file for another file of that type. The file must be named "beep.wav" in order for the emulator to properly pick it up.

The next planned addition is for SCHIP functionality. This will most likely require a detection method to determine what type of ROM is being loaded, and thus, what methods should be used to communicate with SFML.