It works!
After a few weeks of coding here and there, scrolling, SCHIP drawing and SCHIP functionality as a whole seems to be functioning properly. The scaling is currently off, and the projects code as a whole is really, really ugly. It's full of sphagetti code, and weird naming schemes, all proofs of a novice coder at work.

As such, after fixing the scaling, and making sure I haven't broken anything else, I'll merge this with the main branch, then begin working on restructuring the program. I'll build it from the ground up with abstraction and encapsulation in mind. Once that's completely, I'll start focusing on the debugging/GUI portion as this will be an important step to learn for future software development

HURRAY! (changelog below)
-Chip8.cpp
1. I wanted to keep DXYN intact and use it to handle drawing for both SCHIP and CHIP8 sprites. Unfortunately, the combined method is large, unwieldly and ugly. But it works. In the future during the restructure, I may instead split this into two drawing routines, one specifically meant for the other. There's tons of chances for improvement here, but the way it is, it's terrible and needs to be rewritten. I may also choose to forgo using 2D arrays because syntax poison, and instead switching over to a 1D array and just using offsets instead.
2. SCR and SCL both had to be rewritten in certain sections. Both also had to be bounds checked (I did this using the '%' operator which takes the remainder of the division operation of the index and the total range). Once this was completely, I had to implement a type of erasure for columns to give the appearance of a scrolling screen. SCL can possibly be optimized by going "backwards" through the array. This would allow me to avoid having to go through the array three times.