Scaling is fixed.
The float precision loss was due to the divison of ints. At least one variable on one side of the operation has to be cast as a float. Once this was done, the precision was maintained. I can now start looking merging this branch with the main one, and then working on rebuilding the entire project.