A New Beginning
After taking about a month break to focus on work and step back from my code, I've returned and started on my project again.

I've decided that a refactor is in order as my current codebase looks like a jumble. I'll be working on really flexing my class knowledge so I can properly implement encapsulation and abstraction concepts. I'll also begin converting most of my fundamental types to cstdint equivalents in order to get into the habit of doing so, especially with emulator projects, or projects I plan on making cross-platform compatible. .

Last but certainly not least, I'll be taking a look at the jump table I use for my Chip8 opcodes and possibly switching up to a std::map layout instead (if I decide to go this route, it'll be much later down the line once everything else is prettied up).

Now, the changelog can be found below:

New Class:
-EmuCore
1. EmuCore will be the binding class that interacts between Chip and any other external classes or files. This will allow me to pull all of the loose functions in the main.cpp file to clean it up. It'll also make adding additional functionality in the future much easier as I'll have a solid platform to work from.
2. Began fleshing out both the header and cpp file.

Project Changes:
-C Standard Ints
1. I've swapped all fundamental types (that I can) to uint types to standardize type sizes. This helps with portability and explictiness. Some types have remained unchanged, such a bools or floats as they're acted on based on their types and not necessarily their sizes. For instance, I left the char* for the main arguments alone, which means I had to leave the romBuffer type alone as well, otherwise it wouldn't be compatible.

There will be many more changes coming in the future as this is just the start.