Small Update A lot of my work being done currently is research. I'm looking a new projects, especially emulator based ones, learning how they abstract and encapsulate their code, learning about different techniques, libraries and functions.
My current plan is to continue splitting up the code base into core components (display, input handling, etc...), as this seems to be the main way most people keep their code clean. I've read that you don't need a class for every functionality present in your project, but I also have learned that one of the core concepts of C++ and OOP as a whole is "no way is a bad way so long as it runs and runs well". If it's a fun project, do it how you want, if you're cooperating, do it how the team has requested, and stick to that coding standard.

Time for the small notes.

-EmuCore Class:
1. I've removed all parameters from functions pushBuffer, keyStateReleased, keyStatePressed and EmulationLoop. Since the objects being used/passed as parameters were created in the class constructor, there was no reason to pass them around.